refactor: move memory reranking into engine
